Move inner class RecoveredSchedulerState to upper level
