+++ scheduler/scheduler-server/src/test/java/org/ow2/proactive/scheduler/core/db/schedulerdb/BaseSchedulerDBTest.java
@@ -35,8 +35,8 @@ import org.ow2.proactive.scheduler.common.task.TaskResult;
 import org.ow2.proactive.scheduler.common.task.TaskState;
 import org.ow2.proactive.scheduler.common.task.TaskStatus;
 import org.ow2.proactive.scheduler.common.task.executable.JavaExecutable;
+import org.ow2.proactive.scheduler.core.db.RecoveredSchedulerState;
 import org.ow2.proactive.scheduler.core.db.SchedulerDBManager;
-import org.ow2.proactive.scheduler.core.db.SchedulerStateRecoverHelper;
 import org.ow2.proactive.scheduler.core.properties.PASchedulerProperties;
 import org.ow2.proactive.scheduler.descriptor.EligibleTaskDescriptor;
 import org.ow2.proactive.scheduler.job.InternalJob;
@@ -158,8 +158,8 @@ public class BaseSchedulerDBTest extends ProActiveTest {
 		return defaultCredentials;
 	}
 
-	protected SchedulerStateRecoverHelper.RecoveredSchedulerState checkRecoveredState(
-			SchedulerStateRecoverHelper.RecoveredSchedulerState state, StateMatcher stateMatcher) {
+	protected RecoveredSchedulerState checkRecoveredState(
+			RecoveredSchedulerState state, StateMatcher stateMatcher) {
 		assertThat(state.getSchedulerState(), stateMatcher);
 
 		assertThat(state.getPendingJobs(), containsInAnyOrder(stateMatcher.pending.toArray(new JobStateMatcher[] {})));
