+++ scheduler/scheduler-server/src/test/java/org/ow2/proactive/scheduler/core/db/schedulerdb/TestRestoreWorkflowJobs2.java
@@ -5,6 +5,7 @@ import org.ow2.proactive.scheduler.common.task.JavaTask;
 import org.ow2.proactive.scheduler.common.task.flow.FlowAction;
 import org.ow2.proactive.scheduler.common.task.flow.FlowActionType;
 import org.ow2.proactive.scheduler.common.task.flow.FlowScript;
+import org.ow2.proactive.scheduler.core.db.RecoveredSchedulerState;
 import org.ow2.proactive.scheduler.core.db.SchedulerStateRecoverHelper;
 import org.ow2.proactive.scheduler.job.ChangedTasksInfo;
 import org.ow2.proactive.scheduler.job.InternalJob;
@@ -38,7 +39,7 @@ public class TestRestoreWorkflowJobs2 extends BaseSchedulerDBTest {
         dbManager.updateAfterWorkflowTaskFinished(job, changesInfo, result);
 
         SchedulerStateRecoverHelper recoverHelper = new SchedulerStateRecoverHelper(dbManager);
-        SchedulerStateRecoverHelper.RecoveredSchedulerState state = recoverHelper.recover(-1);
+        RecoveredSchedulerState state = recoverHelper.recover(-1);
         job = state.getRunningJobs().get(0);
         System.out.println("OK");
     }
